+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/windmill/work/processing/StreamingWorkScheduler.java
@@ -242,7 +242,6 @@ public class StreamingWorkScheduler {
     long processingStartTimeNanos = System.nanoTime();
     StageInfo stageInfo = getStageInfo(computationState);
 
-    @Nullable List<Work> workBatch = null;
     try {
       if (work.isFailed()) {
         throw new WorkItemCancelledException(workItem.getShardingKey());
@@ -251,7 +250,7 @@ public class StreamingWorkScheduler {
       // Execute the user code for the Work batch.
       ExecuteWorkResult executeWorkResult =
           executeWork(work, stageInfo, computationState, handle, keyTransitionListener);
-      workBatch = executeWorkResult.workBatch();
+      List<Work> workBatch = handle.getWorkBatch();
       List<Windmill.WorkItemCommitRequest> workItemCommits = executeWorkResult.workItemCommits();
 
       commitFinalizer.cacheCommitFinalizers(executeWorkResult.finalizationCallbacks());
@@ -264,7 +263,7 @@ public class StreamingWorkScheduler {
       handleProcessWorkFailure(
           computationState, handle.getWorkBatch(), computationId, systemName, work, t);
     } finally {
-      List<Work> processedWorkBatch = workBatch != null ? workBatch : ImmutableList.of(work);
+      List<Work> processedWorkBatch = handle.getWorkBatch();
       // Update total processing time counters. Updating in finally clause ensures that
       // work items causing exceptions are also accounted in time spent.
       recordProcessingTime(stageInfo, processedWorkBatch, processingStartTimeNanos);
@@ -328,7 +327,6 @@ public class StreamingWorkScheduler {
       computationWorkExecutor.executeWork(
           work, workExecutor, handle, keyTransitionListener, onFailedWorkHandler);
 
-      List<Work> workBatch;
       List<Windmill.WorkItemCommitRequest> workItemCommits;
       Map<Long, Pair<Instant, Runnable>> finalizationCallbacks;
       long stateBytesRead;
@@ -338,9 +336,6 @@ public class StreamingWorkScheduler {
         }
         context.flushState();
 
-        // Retrieve executed works, work item commits, and accumulated callbacks from execution
-        // context
-        workBatch = context.getExecutedWorks();
         workItemCommits = context.getWorkItemCommits();
         finalizationCallbacks = context.getFinalizationCallbacks();
         stateBytesRead = context.getStateBytesRead();
@@ -351,8 +346,7 @@ public class StreamingWorkScheduler {
       computationState.releaseComputationWorkExecutor(computationWorkExecutor);
       computationWorkExecutor = null;
 
-      return ExecuteWorkResult.create(
-          workBatch, workItemCommits, finalizationCallbacks, stateBytesRead);
+      return ExecuteWorkResult.create(workItemCommits, finalizationCallbacks, stateBytesRead);
     } catch (Throwable t) {
       if (computationWorkExecutor != null) {
         // If processing failed due to a thrown exception, close the executionState. Do not
@@ -419,10 +413,6 @@ public class StreamingWorkScheduler {
     }
     for (int i = 0; i < workBatch.size(); i++) {
       Windmill.WorkItemCommitRequest commit = workItemCommits.get(i);
-      // TODO: Retry on commit truncations
-      checkState(
-          !commit.getExceedsMaxWorkItemCommitBytes(),
-          "Commit truncation with multikey bundles not implemented");
       Work w = workBatch.get(i);
       multiKeyBuilder.addRequests(
           commit
@@ -523,15 +513,12 @@ public class StreamingWorkScheduler {
   @AutoValue
   abstract static class ExecuteWorkResult {
     static ExecuteWorkResult create(
-        List<Work> workBatch,
         List<Windmill.WorkItemCommitRequest> workItemCommits,
         Map<Long, Pair<Instant, Runnable>> finalizationCallbacks,
         long stateBytesRead) {
       return new AutoValue_StreamingWorkScheduler_ExecuteWorkResult(
-          workBatch, workItemCommits, finalizationCallbacks, stateBytesRead);
+          workItemCommits, finalizationCallbacks, stateBytesRead);
     }
-
-    abstract List<Work> workBatch();
 
     abstract List<Windmill.WorkItemCommitRequest> workItemCommits();